stats refactored.
